Make Most Warps Faster
Game Freak intentionally slowed down warps between maps with different music by artificially lengthening the fadeout time for the music. The map doesn't fade in from black until this time expires.

This commit sets it to the minimum possible time, making warps much faster.

Additionally, restored the nurse checking the Union Room for nearby players. The check for this is hidden in her heal animation. This avoids the lag inherent with entering the Pokemon Center and puts the check right before it is needed.

+++ include/overworld.h
@@ -121,7 +121,7 @@ extern u8 gGlobalFieldTintMode;
 
 extern bool8 (* gFieldCallback2)(void);
 
-void SetLastHealLocationWarp(u8 healLocaionId);
+void SetLastHealLocationWarp(u32 healLocaionId);
 void LoadMapFromCameraTransition(u8 mapGroup, u8 mapNum);
 void CB2_ReturnToFieldFromDiploma(void);
 void CB2_OverworldBasic(void);
